Separated components and handled error when no image uploaded for the product
